Some fixes for dart2wasm (#38167)
* A few fixes to get things compiling via dart2wasm.

* Use `futureToPromise`.

* Some more fixes for dart2wasm.

* Allow promises to have nullable objects.

+++ lib/web_ui/lib/src/engine/navigation/history.dart
@@ -145,7 +145,7 @@ class MultiEntriesBrowserHistory extends BrowserHistory {
     if (_hasSerialCount(currentState)) {
       final Map<dynamic, dynamic> stateMap =
           currentState! as Map<dynamic, dynamic>;
-      return stateMap['serialCount'] as int;
+      return (stateMap['serialCount'] as double).toInt();
     }
     return 0;
   }
